Address comments and fix a broken script.
