+++ scripts/config-template.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-dss_url=test
-lira_url=test
+dss_url="https://dss.dev.data.humancellatlas.org/v1/"
+lira_url="https://pipelines.dev.data.humancellatlas.org/notifications"
 key_file=test
 lira_secret=test
 google_project=test
