+++ scripts/requirements.txt
@@ -1,3 +1,3 @@
-cryptography
-PyJWT
-requests
+cryptography====2.3.1
+PyJWT==1.6.4
+requests==2.18.4
