+++ scripts/subscribe.py
@@ -1,16 +1,15 @@
 #!/usr/bin/env python
 import sys
-import time
 
 import argparse
-import base64
-import functools
 import json
-import jwt
 import requests
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import rsa
-from datetime import datetime
+
+try:
+    from DCPAuthClient import DCPAuthClient
+except ModuleNotFoundError:
+    print('Cannot find the DCPAuthClient within this folder!')
+    exit(1)
 
 
 def main(arguments=None):
@@ -23,99 +22,10 @@ def main(arguments=None):
 
     if command == 'create':
         create_subscription(args)
-    elif command == 'list':
-        list_subscription(args)
+    elif command == 'get':
+        get_subscriptions(args)
     elif command == 'delete':
         delete_subscription(args)
-
-
-class DCPAuthClient(object):
-    audience = "https://dev.data.humancellatlas.org/"
-
-    def __init__(self, path_to_json_key, trusted_google_project):
-        self.path_to_json_key = path_to_json_key
-        self.trusted_google_project = trusted_google_project
-        self.__token = None
-        # TODO: check the liveness of the token and make it as singleton if applicable
-        self.issue_time = None
-        self.expire_time = None
-
-    def get_auth_header(self):
-        return {'Authorization': 'Bearer {}'.format(self.token)}
-
-    @property
-    def token(self):
-        credentials = DCPAuthClient._from_json(self.path_to_json_key)
-        tok = DCPAuthClient.get_service_jwt(service_credentials=credentials, audience=self.audience)
-        self.verify_jwt(tok, audience=self.audience, trusted_google_project=self.trusted_google_project)
-        return self.__token
-
-    @staticmethod
-    def _from_json(file_path):
-        with open(file_path, 'r') as f:
-            return json.load(f)
-
-    @staticmethod
-    @functools.lru_cache(maxsize=32)
-    def get_openid_config(openid_provider):
-        res = requests.get("https://{}/.well-known/openid-configuration".format(openid_provider))
-        res.raise_for_status()
-        return res.json()
-
-    @staticmethod
-    def get_jwks_uri(openid_provider):
-        if openid_provider.endswith("iam.gserviceaccount.com"):
-            return "https://www.googleapis.com/service_accounts/v1/jwk/{}".format(openid_provider)
-        else:
-            return DCPAuthClient.get_openid_config(openid_provider)["jwks_uri"]
-
-    @staticmethod
-    @functools.lru_cache(maxsize=32)
-    def get_public_keys(openid_provider):
-        keys = requests.get(DCPAuthClient.get_jwks_uri(openid_provider)).json()["keys"]
-        return {
-            key["kid"]: rsa.RSAPublicNumbers(
-                    e=int.from_bytes(base64.urlsafe_b64decode(key["e"] + "==="), byteorder="big"),
-                    n=int.from_bytes(base64.urlsafe_b64decode(key["n"] + "==="), byteorder="big")
-            ).public_key(backend=default_backend())
-            for key in keys
-        }
-
-    @staticmethod
-    def get_service_jwt(service_credentials, audience):
-        iat = time.time()
-        exp = iat + 3600
-        payload = {'iss': service_credentials["client_email"],
-                   'sub': service_credentials["client_email"],
-                   'aud': audience,
-                   'iat': iat,
-                   'exp': exp,
-                   'https://auth.data.humancellatlas.org/email': service_credentials["client_email"],
-                   'https://auth.data.humancellatlas.org/group': 'hca',
-                   'scope': ["openid", "email", "offline_access"]
-                   }
-        additional_headers = {'kid': service_credentials["private_key_id"]}
-        signed_jwt = jwt.encode(payload, service_credentials["private_key"], headers=additional_headers,
-                                algorithm='RS256').decode()
-        return signed_jwt
-
-    def verify_jwt(self, token, audience, trusted_google_project):
-        assert token
-        try:
-            openid_provider = "humancellatlas.auth0.com"
-            token_header = jwt.get_unverified_header(token)
-            public_keys = DCPAuthClient.get_public_keys(openid_provider)
-            tok = jwt.decode(token, key=public_keys[token_header["kid"]], audience=audience)
-        except KeyError:
-            unverified_token = jwt.decode(token, verify=False)
-            issuer = unverified_token["iss"]
-            assert issuer.endswith("@{}.iam.gserviceaccount.com".format(trusted_google_project))
-            token_header = jwt.get_unverified_header(token)
-            public_keys = DCPAuthClient.get_public_keys(issuer)
-            tok = jwt.decode(token, key=public_keys[token_header["kid"]], audience=audience)
-        self.__token = token
-        self.issue_time = datetime.fromtimestamp(tok['iat']).strftime('%Y-%m-%d %H:%M:%S')
-        self.expire_time = datetime.fromtimestamp(tok['exp']).strftime('%Y-%m-%d %H:%M:%S')
 
 
 def create_subscription(args):
@@ -172,7 +82,7 @@ def _prep_json(callback_base_url, query_json_file, query_param_token, hmac_key_i
     return payload
 
 
-def list_subscription(args):
+def get_subscriptions(args):
     print('Listing all subscriptions in replica: {0} of DSS: {1}\n'.format(
             args['replica'],
             args['dss_url']))
@@ -207,10 +117,10 @@ def parser(arguments):
     subparsers = main_parser.add_subparsers(help='All available commands', dest='command')
     create = subparsers.add_parser('create', help='Create a new subscription in DSS.')
     delete = subparsers.add_parser('delete', help='Delete an existing subscription in DSS.')
-    list = subparsers.add_parser('list', help='List all subscriptions in DSS.')
+    get = subparsers.add_parser('get', help='Get and list all subscriptions in DSS.')
 
     # TODO rex: group sub-commands
-    subscribe_sub_commands = (create, delete, list)
+    subscribe_sub_commands = (create, delete, get)
 
     # Add common arguments for all of the commands
     for sub_command in subscribe_sub_commands:
@@ -222,7 +132,6 @@ def parser(arguments):
                                  required=True)
         sub_command.add_argument('--replica',
                                  help='Which replica to work on, use "gcp" by default. ["gcp", "aws"]',
-                                 # required=True,
                                  default='gcp')
         sub_command.add_argument('--google_project',
                                  help='The google project the Lira is using.',
@@ -270,4 +179,4 @@ def parser(arguments):
 
 
 if __name__ == '__main__':
-    main(sys.argv[1:])
+    main(sys.argv[1:])  # exclude the python file name from the input arguments
