+++ scripts/subscription-delete.sh
@@ -2,10 +2,10 @@
 
 # Usage: bash subscription-delete.sh config.sh subscription_id
 
-config_script=$1
+script_config=$1
 subscription_id=$2
 
-source $config_script
+source $script_config
 python3 subscribe.py delete --dss_url="$dss_url" \
                             --key_file="$key_file" \
                             --google_project="$google_project" \
